fix: Compilation issue when using IStorageAPI (should have required importation of Luxoria.Core)
